+++ src/client/components/SearchInput.tsx
@@ -18,26 +18,31 @@ const SearchInput: React.SFC<SearchInputProps> = (props: SearchInputProps) => {
   return (
     <div className="search__container__outer">
       <div className="search__container__inner">
-        <div className="input__addon__left">
-          <i className="material-icons grey icon-sm">work_outline</i>
-        </div>
-        <input
-          aria-label="Title, companies, expertise or benefits"
-          className="search__input"
-          id="search"
-          onChange={(e) => setSearch(e.target.value)}
-          placeholder="Title, companies, expertise or benefits"
-          type="text"
-          value={search}
-        />
-        <div className="search__button__container">
-          <button
-            className="search__button"
-            onClick={() => handleSearch(search, locationOptions)}
-          >
-            Search
-          </button>
-        </div>
+        <form
+          id="search-form"
+          onSubmit={(e) => {
+            e.preventDefault();
+            handleSearch(search, locationOptions);
+          }}
+        >
+          <div className="input__addon__left">
+            <i className="material-icons grey icon-sm">work_outline</i>
+          </div>
+          <input
+            aria-label="Title, companies, expertise or benefits"
+            className="search__input"
+            id="search"
+            onChange={(e) => setSearch(e.target.value)}
+            placeholder="Title, companies, expertise or benefits"
+            type="text"
+            value={search}
+          />
+          <div className="search__button__container">
+            <button className="search__button" type="submit">
+              Search
+            </button>
+          </div>
+        </form>
       </div>
     </div>
   );