+++ cypress/integration/search.spec.js
@@ -27,4 +27,44 @@ context("Search", () => {
       assert.equal($jobs.length, 5);
     });
   });
+
+  it("Should retain search value on reload", () => {
+    cy.get("#search").type("developer");
+    cy.get(".search__button").click();
+    cy.wait(1000);
+    cy.reload();
+    cy.get("#search").should("have.value", "developer");
+  });
+
+  it("Should be able to submit form with enter key", () => {
+    cy.get(".orbit-spinner").should("not.be.visible");
+    cy.get("#search").type("developer");
+    cy.get("#search").type("{enter}");
+    cy.get(".orbit-spinner").should("be.visible");
+  });
+});
+
+context("Search - No Results", () => {
+  beforeEach(() => {
+    cy.server();
+    cy.route({
+      method: "GET",
+      url: "/jobs",
+      status: 200,
+      response: [],
+      onRequest: (xhr) => {},
+      onResponse: (xhr) => {},
+    });
+    cy.visit("http://localhost:3000");
+  });
+
+  it("Should display note when there are no results", () => {
+    cy.get("#search").type("developer");
+    cy.get(".search__button").click();
+    cy.wait(1000);
+    cy.get("#no-results").should(
+      "have.text",
+      "No results. Please modify your search and try again."
+    );
+  });
 });