+++ src/server/controllers/assets.ts
@@ -12,12 +12,9 @@ class AssetsController {
     "android-chrome-512x512.png",
     "apple-touch-icon.png",
     "backgroundImg.png",
-    "browserconfig.xml",
+    "favicon.ico",
     "favicon-16x16.png",
     "favicon-32x32.png",
-    "favicon.ico",
-    "mstile-150x150.png",
-    "safari-pinned-tab.svg",
     "site.webmanifest",
   ];