+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - If "How To Apply" is a single link, display a styled link
+- Page title
+- Page description
 
 ### Removed